# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareAI Healthcare Triage API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile and preferences
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/preferences:
    patch:
      operationId: updateUserPreferences
      tags: [user]
      summary: Update user preferences (language, voice settings)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPreferencesInput"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /telemedicine/doctors:
    get:
      operationId: listDoctors
      tags: [telemedicine]
      summary: List available doctors
      parameters:
        - name: specialty
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Doctors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /telemedicine/book:
    post:
      operationId: bookAppointment
      tags: [telemedicine]
      summary: Book a telemedicine appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      required: [id, email, name, language, voiceEnabled, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        language:
          type: string
          enum: [en, hi, kn]
        voiceEnabled:
          type: boolean
        voiceRate:
          type: number
        createdAt:
          type: string
          format: date-time
    UserPreferencesInput:
      type: object
      properties:
        language:
          type: string
          enum: [en, hi, kn]
        voiceEnabled:
          type: boolean
        voiceRate:
          type: number
...
    Doctor:
      type: object
      required: [id, name, specialty, experience, rating, reviewCount, available]
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        experience:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        available:
          type: boolean
        avatarEmoji:
          type: string
        nextSlot:
          type: ["string", "null"]
    AppointmentInput:
      type: object
      required: [doctorId, type]
      properties:
        doctorId:
          type: integer
        type:
          type: string
          enum: [video, chat, callback]
        userId:
          type: ["integer", "null"]
        sessionId:
          type: ["string", "null"]
        preferredTime:
          type: ["string", "null"]
    Appointment:
      type: object
      required: [id, doctorId, type, status, createdAt]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        type:
          type: string
        status:
          type: string
        doctor:
          $ref: "#/components/schemas/Doctor"
        scheduledAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...